requirements.txt
argon2-cffi==20.1.0
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
